Convert SilkTrader plugin from kts to kt
